+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 #### Android
 - Use MediaStore Opener (which goes through the gallery) instead of default explorer. (Thank you @tmthecoder).
 
+#### Web
+- Add event when canceling the picker. (Thank you @letranloc).
+
 ## 3.0.0
 Adds null safety support ([#510](https://github.com/miguelpruivo/flutter_file_picker/issues/510)).
 ## 2.1.7